refactor(backend): split routers and dtos in files
